(fix) java.StackOverflow on PostController

+++ src/main/java/com/stormx/hicoder/services/implement/PostServiceImpl.java
@@ -1,6 +1,7 @@
 package com.stormx.hicoder.services.implement;
 
 import com.stormx.hicoder.controllers.requests.NewPostRequest;
+import com.stormx.hicoder.dto.PostDTO;
 import com.stormx.hicoder.entities.Post;
 import com.stormx.hicoder.entities.User;
 import com.stormx.hicoder.exceptions.BadRequestException;
@@ -19,8 +20,11 @@ public class PostServiceImpl implements PostService {
     private final UserRepository userRepository;
 
     @Override
-    public List<Post> getAllPostsOfUser(User user) {
-        return postRepository.findAllByAuthor(user);
+    public List<PostDTO> getAllPostsOfUser(User user) {
+        List<Post> userPosts = postRepository.findAllByAuthor(user);
+        return userPosts.stream()
+                .map(PostDTO::new)
+                .toList();
     }
 
     @Override
@@ -41,20 +45,21 @@ public class PostServiceImpl implements PostService {
     @Override
     public void updatePost(String postId, NewPostRequest postDetails, User currentUser) {
         Post oldPost = getPostById(postId);
-//        if(!currentUser.isPostedBy(oldPost)){
-//            throw new BadRequestException("User doesn't have post: " + postId);
-//        }
+        if (!oldPost.isPostedBy(currentUser)) {
+            throw new BadRequestException("User doesn't have post: " + postId);
+        }
         postRepository.save(oldPost);
     }
 
     @Override
     public void deletePost(String postId, User currentUser) {
         Post post = getPostById(postId);
-//        if(!currentUser.isPostedBy(post)){
-//            throw new BadRequestException("User doesn't have post: " + postId);
-//        }
+        if (!post.isPostedBy(currentUser)) {
+            throw new BadRequestException("User doesn't have post: " + postId);
+        }
+        currentUser.removePost(post);
+        userRepository.save(currentUser);
         postRepository.delete(post);
-//        currentUser.removePost(post);
     }
 
 }
